move data
remove unneeded

try fix

Disable `test-tutorials-cratesio` (#4970)

We're now making breaking changes, so we cannot keep this test.

try fix

revert removal

revert removal

fix
