+++ .github/workflows/build-test.yml
@@ -144,7 +144,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix: 
-        behavior: [local, cratesio]
+        behavior: [local] # reenable cratesio for 2.0
       fail-fast: false
     steps:
     - uses: actions/checkout@v4
