+++ components/experimental/src/dimension/currency/mod.rs
@@ -7,3 +7,4 @@
 pub mod format;
 pub mod formatter;
 pub mod options;
+pub mod provider;
