+++ components/experimental/src/dimension/currency/provider/currency.rs
@@ -38,7 +38,7 @@ pub use crate::provider::Baked;
 #[cfg_attr(
     feature = "datagen",
     derive(serde::Serialize, databake::Bake),
-    databake(path = icu_experimental::dimension::provider::currency),
+    databake(path = icu_experimental::dimension::currency::provider::currency),
 )]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize))]
 #[yoke(prove_covariance_manually)]
@@ -76,7 +76,7 @@ pub struct CurrencyEssentialsV1<'data> {
 #[cfg_attr(
     feature = "datagen",
     derive(serde::Serialize, databake::Bake),
-    databake(path = icu_experimental::dimension::provider::currency),
+    databake(path = icu_experimental::dimension::currency::provider::currency),
 )]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize))]
 #[derive(Copy, Clone, Debug, PartialOrd, Ord, PartialEq, Eq, Default)]
@@ -93,7 +93,7 @@ pub enum PatternSelection {
 #[cfg_attr(
     feature = "datagen",
     derive(serde::Serialize, databake::Bake),
-    databake(path = icu_experimental::dimension::provider::currency),
+    databake(path = icu_experimental::dimension::currency::provider::currency),
 )]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize))]
 #[derive(Copy, Debug, Clone, PartialEq, PartialOrd, Eq, Ord)]
@@ -109,7 +109,7 @@ pub enum PlaceholderValue {
 #[cfg_attr(
     feature = "datagen",
     derive(serde::Serialize, databake::Bake),
-    databake(path = icu_experimental::dimension::provider::currency),
+    databake(path = icu_experimental::dimension::currency::provider::currency),
 )]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize))]
 #[derive(Copy, Debug, Clone, Default, PartialEq, PartialOrd, Eq, Ord)]
