+++ components/experimental/src/lib.rs
@@ -88,8 +88,8 @@ pub mod provider {
         super::compactdecimal::provider::ShortCompactDecimalFormatDataV1Marker::KEY,
         super::compactdecimal::provider::LongCompactDecimalFormatDataV1Marker::KEY,
         super::compactdecimal::provider::ShortCompactDecimalFormatDataV1Marker::KEY,
-        super::dimension::provider::currency::CurrencyEssentialsV1Marker::KEY,
-        super::dimension::provider::percent::PercentEssentialsV1Marker::KEY,
+        super::dimension::currency::provider::currency::CurrencyEssentialsV1Marker::KEY,
+        super::dimension::currency::provider::percent::PercentEssentialsV1Marker::KEY,
         super::displaynames::provider::LanguageDisplayNamesV1Marker::KEY,
         super::displaynames::provider::LocaleDisplayNamesV1Marker::KEY,
         super::displaynames::provider::RegionDisplayNamesV1Marker::KEY,
