+++ provider/datagen/src/registry.rs
@@ -287,8 +287,8 @@ macro_rules! registry(
             icu::datetime::provider::neo::TimeNeoSkeletonPatternsV1Marker = "datetime/patterns/time_skeleton@1",
             icu::experimental::compactdecimal::provider::LongCompactDecimalFormatDataV1Marker = "compactdecimal/long@1",
             icu::experimental::compactdecimal::provider::ShortCompactDecimalFormatDataV1Marker = "compactdecimal/short@1",
-            icu::experimental::dimension::provider::currency::CurrencyEssentialsV1Marker = "currency/essentials@1",
-            icu::experimental::dimension::provider::percent::PercentEssentialsV1Marker = "percent/essentials@1",
+            icu::experimental::dimension::currency::provider::currency::CurrencyEssentialsV1Marker = "currency/essentials@1",
+            icu::experimental::dimension::currency::provider::percent::PercentEssentialsV1Marker = "percent/essentials@1",
             icu::experimental::displaynames::provider::RegionDisplayNamesV1Marker = "displaynames/regions@1",
             icu::experimental::displaynames::provider::LanguageDisplayNamesV1Marker = "displaynames/languages@1",
             icu::experimental::displaynames::provider::LocaleDisplayNamesV1Marker = "displaynames/locales@1",
